OCSADV-200: move everything new into a new package (temporarily at least)
